fix(import): extract config-tab helper and chunk row deletions
Two bugs fixed:
1. Importing upsertConfigValue from setup-sheet.ts triggered its unguarded
   main(), running full sheet setup on every import. Moved upsertConfigValue
   to a new scripts/config-tab.ts; setup-sheet.ts re-exports it. Both import
   scripts now import directly from config-tab.ts.

2. deleteTransactionsBeforeCutover sent all deleteDimension requests in a
   single batchUpdate, hitting Google Sheets API payload limits on large
   datasets (~26k rows). Chunked into batches of 500 with progress logging.
